requirements.txt
streamlit>=1.32
pandas>=2.1
plotly>=5.18
streamlit-aggrid>=0.3.4
pillow>=10.1
click>=8.1
pytest>=7.4
numpy>=1.26
